Add the connector.cpp and terminate.cpp files
This commit will add the source files for the Connector and Terminate
classes. Connector is an abstract class used for common data members
between the connector classes. Terminate executes the left and right
commands and returns the result of the right one, but this may later
be modified to always return true.

Closes #6